+++ CHANGES.rst
@@ -12,7 +12,10 @@ Changes:
 --------
 - Add `Vault` endpoints providing a secured self-hosted file storage to upload local files for execution input.
 - Add ``upload`` CLI operation for uploading local files to `Vault`.
+- Add CLI automatic detection of local files during ``execute`` call to upload to `Vault` and retrieve them from it
+  on the remote `Weaver` instance.
 - Improve Media-Type/Content-Type guesses based on known local definitions and extensions in ``weaver.formats``.
+- Extend ``PyWPS`` ``WPSRequest`` to support more authorization header forwarding for inputs that could need it.
 
 Fixes:
 ------
@@ -21,6 +24,7 @@ Fixes:
 - Fix rendering of OpenAPI variable names (``additionalParameters``) employed to represent for example ``{input-id}``
   as the key within the mapping representation of inputs/outputs. The previous notation employed was incorrectly
   interpreted as HTML tags, making them partially hidden in Swagger UI.
+- Fix invalid generation of timezone-aware datetime.
 
 `4.8.0 <https://github.com/crim-ca/weaver/tree/4.8.0>`_ (2022-01-11)
 ========================================================================